Rename persistence package
