+++ Patient-Service/src/main/java/com/arivanamin/healthcare/backend/patient/details/JpaPatientPersistence.java
@@ -1,4 +1,4 @@
-package com.arivanamin.healthcare.backend.patient.persistence;
+package com.arivanamin.healthcare.backend.patient.details;
 
 import com.arivanamin.healthcare.backend.patient.core.entity.Patient;
 import com.arivanamin.healthcare.backend.patient.core.persistence.PatientPersistence;
