+++ Patient-Service/src/main/java/com/arivanamin/healthcare/backend/patient/details/PatientRepository.java
@@ -1,4 +1,4 @@
-package com.arivanamin.healthcare.backend.patient.persistence;
+package com.arivanamin.healthcare.backend.patient.details;
 
 import org.springframework.data.jpa.repository.JpaRepository;
